+++ src/main/java/com/transport/HelperApiApplication.java
@@ -1,4 +1,4 @@
-package com.transport.helper_api;
+package com.transport;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
@@ -9,5 +9,4 @@ public class HelperApiApplication {
 	public static void main(String[] args) {
 		SpringApplication.run(HelperApiApplication.class, args);
 	}
-
 }
