+++ src/main/java/com/transport/model/Location.java
@@ -1,5 +1,6 @@
-package com.transport.helper_api.model;
+package com.transport.model;
 
+import org.springframework.data.neo4j.core.schema.GeneratedValue;
 import org.springframework.data.neo4j.core.schema.Id;
 import org.springframework.data.neo4j.core.schema.Node;
 
@@ -13,6 +14,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 public class Location {
 	@Id
+	@GeneratedValue
 	private Long id;
 	private String name;
 	private double longitude;
