+++ src/main/java/com/transport/model/TestNode.java
@@ -1,4 +1,4 @@
-package com.transport.helper_api.model;
+package com.transport.model;
 
 import lombok.Data;
 import org.springframework.data.neo4j.core.schema.GeneratedValue;
